# backend/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/tracks/favourites": {
      "get": {
        "summary": "Get Favourite Tracks",
        "operationId": "get_favourite_tracks_tracks_favourites_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTracksListSchema"
                }
              }
            }
          }
        }
      }
    },
    "/tracks/{track_id}/similar": {
      "get": {
        "summary": "Get Similar Tracks",
        "operationId": "get_similar_tracks_tracks__track_id__similar_get",
        "parameters": [
          {
            "name": "track_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Track Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTracksListSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tracks/{track_id}/like": {
      "post": {
        "summary": "Post Like Track",
        "operationId": "post_like_track_tracks__track_id__like_post",
        "parameters": [
          {
            "name": "track_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Track Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/peaks/{track_id}": {
      "get": {
        "summary": "Get Peak",
        "operationId": "get_peak_peaks__track_id__get",
        "parameters": [
          {
            "name": "track_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Track Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetPeakSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/playlists/my": {
      "get": {
        "summary": "Get My Playlists",
        "operationId": "get_my_playlists_playlists_my_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetPlaylistsListSchema"
                }
              }
            }
          }
        }
      }
    },
    "/playlists/{playlist_id}/tracks": {
      "get": {
        "summary": "Get Playlist Tracks",
        "operationId": "get_playlist_tracks_playlists__playlist_id__tracks_get",
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Playlist Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTracksListSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/playlists": {
      "get": {
        "summary": "Search Playlists",
        "operationId": "search_playlists_playlists_get",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetPlaylistsListSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/artists/{artist_id}/tracks": {
      "get": {
        "summary": "Get Artist Tracks",
        "operationId": "get_artist_tracks_artists__artist_id__tracks_get",
        "parameters": [
          {
            "name": "artist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artist Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTracksListSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/artists": {
      "get": {
        "summary": "Search Artists",
        "operationId": "search_artists_artists_get",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetArtistsListSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "GetArtistSchema": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "cover_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "cover_url"
        ],
        "title": "GetArtistSchema"
      },
      "GetArtistsListSchema": {
        "properties": {
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "artists": {
            "items": {
              "$ref": "#/components/schemas/GetArtistSchema"
            },
            "type": "array",
            "title": "Artists"
          }
        },
        "type": "object",
        "required": [
          "count",
          "artists"
        ],
        "title": "GetArtistsListSchema"
      },
      "GetPeakSchema": {
        "properties": {
          "track_id": {
            "type": "string",
            "title": "Track Id"
          },
          "download_url": {
            "type": "string",
            "title": "Download Url"
          }
        },
        "type": "object",
        "required": [
          "track_id",
          "download_url"
        ],
        "title": "GetPeakSchema"
      },
      "GetPlaylistSchema": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "cover_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "cover_url"
        ],
        "title": "GetPlaylistSchema"
      },
      "GetPlaylistsListSchema": {
        "properties": {
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "playlists": {
            "items": {
              "$ref": "#/components/schemas/GetPlaylistSchema"
            },
            "type": "array",
            "title": "Playlists"
          }
        },
        "type": "object",
        "required": [
          "count",
          "playlists"
        ],
        "title": "GetPlaylistsListSchema"
      },
      "GetTrackSchema": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "artists": {
            "items": {
              "$ref": "#/components/schemas/GetArtistSchema"
            },
            "type": "array",
            "title": "Artists"
          },
          "cover_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "artists",
          "cover_url"
        ],
        "title": "GetTrackSchema"
      },
      "GetTracksListSchema": {
        "properties": {
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "tracks": {
            "items": {
              "$ref": "#/components/schemas/GetTrackSchema"
            },
            "type": "array",
            "title": "Tracks"
          }
        },
        "type": "object",
        "required": [
          "count",
          "tracks"
        ],
        "title": "GetTracksListSchema"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}